+++ R/serialize.R
@@ -2,7 +2,7 @@
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 #' Save an encrypted R object to file
 #' 
-#' @inheritParams encrypt
+#' @inheritParams encrypt_raw
 #' @param robj R object
 #' @param filename destination filename for encrypted R object
 #' 
@@ -10,10 +10,10 @@
 #' @export
 #' @examples
 #' file <- tempfile()
-#' encrypt_obj(head(mtcars), file, key = "hello")
-#' decrypt_obj(file, key = "hello")
+#' encrypt(head(mtcars), file, key = "hello")
+#' decrypt(file, key = "hello")
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-encrypt_obj <- function(robj, filename, key) {
+encrypt <- function(robj, filename, key = getOption("MONOCYPHER_KEY", default = NULL)) {
   filename <- normalizePath(filename, mustWork = FALSE)
   .Call(serialize_, robj, filename, key)
   invisible(filename)
@@ -22,17 +22,17 @@ encrypt_obj <- function(robj, filename, key) {
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 #' Read an encrypted R object from file
 #' 
-#' @inheritParams encrypt
+#' @inheritParams encrypt_raw
 #' @param filename source filename containing an encrypted R object
 #'
 #' @return None
 #' @export
 #' @examples
 #' file <- tempfile()
-#' encrypt_obj(head(mtcars), file, key = "hello")
-#' decrypt_obj(file, key = "hello")
+#' encrypt(head(mtcars), file, key = "hello")
+#' decrypt(file, key = "hello")
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-decrypt_obj <- function(filename, key) {
+decrypt <- function(filename, key = getOption("MONOCYPHER_KEY", default = NULL)) {
   filename <- normalizePath(filename)
   .Call(unserialize_, filename, key)
 }